+++ TelegramBot/TelegramBot/Data/Definitions/Role.cs
@@ -1,0 +1,8 @@
+﻿namespace TelegramBot.Data.Definitions
+{
+    public enum Role
+    {
+        USER,
+        ADMIN
+    }
+}
