+++ TelegramBot/TelegramBot/Data/Entities/Day.cs
