+++ TelegramBot/TelegramBot/Data/Entities/Lesson.cs
