+++ TelegramBot/TelegramBot/Data/Entities/LessonsAndTimes.cs
