+++ TelegramBot/TelegramBot/Data/Entities/News.cs
