+++ TelegramBot/TelegramBot/Data/SubModels/RegistrationModel.cs
