+++ TelegramBot/TelegramBot/Messages/MessageCollector.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using TelegramBot.Data;
+using TelegramBot.Data.Definitions;
 using TelegramBot.Data.Models;
 using TelegramBot.Data.ViewModels;
 using TelegramBot.Interfaces;
